gpu_services/embedding_service/requirements.txt
# Gemini Embedding Service Requirements
# No GPU/torch dependencies - uses cloud API

# Gemini SDK
google-genai

# FastAPI framework
fastapi
uvicorn[standard]

# Data validation
pydantic
pydantic-settings

# Retry logic
tenacity

# Async HTTP
aiohttp

# Numerical operations (for normalization)
numpy